% File alkans74.RData
\name{alkans74.mdb}
\docType{data}
\alias{alkans74.mdb}
\title{Molecular database with 74 Alkanes}
\description{
  Molecular database consisting of 74 properties.
}
\usage{alkans74.mdb}
\format{Molecular database - list of molecules with properties}
\source{D.E.Needham, I.C.Wei, P.G.Seybold, J.Am.Chem.Soc., 110, 4186-4194 (1988)}
\keyword{datasets}
