% File cinf-classperf.R
\name{calc_accuracy}
\alias{calc_accuracy}
\title{Calculate classification accuracy}
\description{
  Calculates classification accuracy
}
\usage{
calc_accuracy(tp, tn, fp, fn)
}
\arguments{
	\item{tp}{true positives}
	\item{tn}{true negatives}
	\item{fp}{false positives}
	\item{fn}{false negatives}
}
\value{Classification accuracy}
\examples{
}
