% File cinf-classperf.R
\name{calc_balaccuracy}
\alias{calc_balaccuracy}
\title{Calculate classification balanced accuracy}
\description{
  Calculates classification balanced accuracy
}
\usage{
calc_balaccuracy(tp, tn, fp, fn)
}
\arguments{
	\item{tp}{true positives}
	\item{tn}{true negatives}
	\item{fp}{false positives}
	\item{fn}{false negatives}
}
\value{Classification balanced accuracy}
\examples{
}
