% File cinf-classperf.R
\name{calc_f1}
\alias{calc_f1}
\title{Calculate classification F1-score}
\description{
  Calculates classification F1-score
}
\usage{
calc_f1(tp, tn, fp, fn)
}
\arguments{
	\item{tp}{true positives}
	\item{tn}{true negatives}
	\item{fp}{false positives}
	\item{fn}{false negatives}
}
\value{Classification F1-score}
\examples{
}
