% File cinf-classperf.R
\name{calc_kappa}
\alias{calc_kappa}
\title{Calculate kappa-statistics}
\description{
  Calculates kappa-statistics
}
\usage{
calc_kappa(tp, tn, fp, fn)
}
\arguments{
	\item{tp}{true positives}
	\item{tn}{true negatives}
	\item{fp}{false positives}
	\item{fn}{false negatives}
}
\value{Kappa-statistics}
\examples{
}
