% File cinf-classperf.R
\name{calc_matcorcoef}
\alias{calc_matcorcoef}
\title{Calculate Matthews correlation coefficient}
\description{
  Calculates Matthews correlation coefficient
}
\usage{
calc_matcorcoef(tp, tn, fp, fn)
}
\arguments{
	\item{tp}{true positives}
	\item{tn}{true negatives}
	\item{fp}{false positives}
	\item{fn}{false negatives}
}
\value{Matthews correlation coefficient}
\examples{
}
