% File cinf-classperf.R
\name{calc_precision}
\alias{calc_precision}
\title{Calculate classification precision}
\description{
  Calculates classification precision
}
\usage{
calc_precision(tp, tn, fp, fn)
}
\arguments{
	\item{tp}{true positives}
	\item{tn}{true negatives}
	\item{fp}{false positives}
	\item{fn}{false negatives}
}
\value{Classification precision}
\examples{
}
