% File cinf-classperf.R
\name{calc_recall}
\alias{calc_recall}
\title{Calculate classification recall}
\description{
  Calculates classification recall
}
\usage{
calc_recall(tp, tn, fp, fn)
}
\arguments{
	\item{tp}{true positives}
	\item{tn}{true negatives}
	\item{fp}{false positives}
	\item{fn}{false negatives}
}
\value{Classification recall}
\examples{
}
