% File cinf-classperf.R
\name{calc_sensitivity}
\alias{calc_sensitivity}
\title{Calculate classification sensitivity}
\description{
  Calculates classification sensitivity
}
\usage{
calc_sensitivity(tp, tn, fp, fn)
}
\arguments{
	\item{tp}{true positives}
	\item{tn}{true negatives}
	\item{fp}{false positives}
	\item{fn}{false negatives}
}
\value{Classification sensitivity}
\examples{
}
