% File cinf-classperf.R
\name{calc_specificity}
\alias{calc_specificity}
\title{Calculate classification specificity}
\description{
  Calculates classification specificity
}
\usage{
calc_specificity(tp, tn, fp, fn)
}
\arguments{
	\item{tp}{true positives}
	\item{tn}{true negatives}
	\item{fp}{false positives}
	\item{fn}{false negatives}
}
\value{Classification specificity}
\examples{
}
