% File cinf-free-wilson.R
\name{extract_new}
\alias{extract_new}
\title{Extract rows from new data frame corresponding to compounds not contained in old data frame}
\description{
  Extracts rows from new data frame corresponding to compounds not contained in old data frame
}
\usage{
extract_new(old_df, new_df)
}
\arguments{
	\item{old_df}{old data frame}
	\item{new_df}{new data frame}
}
\value{}
\examples{
}
