% File cinf-free-wilson.R
\name{predict_free_wilson_mlr}
\alias{predict_free_wilson_mlr}
\title{Make predictions for MLR model in the Free-Wilson approach}
\description{
  Makes predictions for MLR model in the Free-Wilson approach.
}
\usage{
predict_free_wilson_mlr(t, model, substlist)
}
\arguments{
	\item{t}{initial table used to build the model}
	\item{model}{MLR model}
	\item{substlist}{list of substitutions, example: list(R1="H", R2="Br")}
}
\value{Predictions}
\examples{
}
