% File cinf-free-wilson.R
\name{predict_table_free_wilson_mlr}
\alias{predict_table_free_wilson_mlr}
\title{Makes predictions for data frame using MLR model in the Free-Wilson approach}
\description{
  Makes predictions for data frame using MLR model in the Free-Wilson approach.
}
\usage{
predict_table_free_wilson_mlr(model, table)
}
\arguments{
	\item{model}{MLR model}
	\item{table}{table with data for prediction}
}
\value{Table with predictions}
\examples{
}
