% File PT.RData
\name{PT}
\docType{data}
\alias{PT}
\title{Periodic Table}
\description{
  Properties of chemical elements.
}
\usage{PT}
\format{
  \tabular{lll}{
    NumEl    \tab numeric \tab The number of electrons for each chemical elements        \cr
	NumValEl \tab numeric \tab The number of valence electrons for each chemical element \cr
	ComVal   \tab numeric \tab The common valence for each chemical element              \cr
	MaxVal   \tab numeric \tab The maximum valence for each chemical element             \cr
	AW       \tab numeric \tab Atomic weight for each chemical element                   \cr
	ElNeg    \tab numeric \tab Electronegativity for each chemical element               \cr
	AtRad    \tab numeric \tab Atomic radius for each chemical element                   \cr
	IonPot   \tab numeric \tab Ionozation potential for each chemical element            \cr
	Color    \tab numeric \tab Color for each chemical element
  }
}
\keyword{datasets}
